Isolate obsolete EnableByDefault tests with pragma suppressions (#7446)
* Initial plan

* Isolate EnableByDefault tests into partial classes with suppressions


* Restore global.json (accidentally renamed in previous commit)


* Collapse When_features_are_scanned into single file with pragma at top


* Add pragma warning restore at bottom of files with pragma disable


---------
